Backend/requirements.txt
# ------------------------
# Django & API
# ------------------------
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-environ>=0.11
django-filter>=24.2

# ------------------------
# Auth & Users
# ------------------------
djangorestframework-simplejwt>=5.3.1
bcrypt>=4.1.2

# ------------------------
# Data / Analytics
# ------------------------
duckdb>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.3
pydantic>=2.7.0
sqlglot>=23.7.0
python-dotenv>=1.0.1
openpyxl>=3.1.2

# ------------------------
# Dev Tools
# ------------------------
black>=24.3.0
flake8>=7.0.0
isort>=5.13.2
pytest>=8.2.0
pytest-django>=4.8.0
coverage>=7.5.1
